mailbridge/requirements.txt
fastapi>=0.115.0
uvicorn==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
bcrypt==4.1.2
google-auth==2.26.2
